make - Faq divide FaqForm and FaqAccordion
