+++ app/[locale]/(root)/(home)/page.tsx
@@ -2,7 +2,7 @@ import Hero from '@/src/components/common/Hero/Hero'
 import TopFlights from '@/src/components/shared/TopFlights/TopFlights'
 import { useTranslations } from 'next-intl'
 import { unstable_setRequestLocale } from 'next-intl/server'
-import FrequentlyAskedQuestions from "@/src/components/shared/FrequentlyAskedQuestions/FrequentlyAskedQuestions";
+import FAQSection from "@/src/components/shared/FAQ-section/FAQ-section";
 import {SpecialDeals} from "@/src/components/shared/SpecialDeals/SpecialDeals";
 import AirlinesRating from "@/src/components/shared/AirlinesRating/AirlinesRating";
 import { Locale } from '@/src/shared/types/i18n';
@@ -15,9 +15,9 @@ export default function Home({ params }: { params: { locale: Locale } }) {
 		<>
 			<Hero title={t('home-title')} preTitle={t('home-desc')} />
 			<TopFlights />
-			<SpecialDeals/>
 			<AirlinesRating/>
-			<FrequentlyAskedQuestions/>
+			<SpecialDeals/>
+			<FAQSection/>
 		</>
 	)
 }
