+++ src/shared/types/faqFormType.ts
@@ -1,0 +1,5 @@
+export interface faqForm  {
+    name?:string,
+    email:string,
+    message:string
+}
