+++ translations/ky.json
@@ -22,7 +22,8 @@
 			"input_area": "Бизге суроо бериңиз",
 			"button": "Аткаруу",
 			"under_button": "Мен Купуялык саясатынын шарттарын кабыл алам жана жеке маалыматтарды иштетүүгө макулмун"
-		}
+		},
+		"loading": "Жүктөлүү..."
 	},
 	"SpecialDeals": {
 		"title": "Атайын сунуштар",
