+++ translations/ru.json
@@ -22,7 +22,8 @@
 			"input_area": "Задайте нам вопрос",
 			"button": "Выполнить",
 			"under_button": "Я принимаю условия Политики конфиденциальности и даю согласие на обработку персональных данных"
-		}
+		},
+		"loading": "Загрузка..."
 	},
 	"SpecialDeals": {
 		"title": "Специальные предложения",
